Integrated to plugin environment.
