--- novacraft-plugin-suite/src/main/java/io/novacraft/bloodmoon/BloodMoonListener.java
@@ -1,9 +1,0 @@
-import org.bukkit.event.Listener;
-
-public class BloodMoonListener implements Listener {
-	
-	//Constructor
-	public BloodMoonListener(BloodMoon plugin) {
-		
-	}
-}
