+++ novacraft-plugin-suite/src/main/java/io/novacraft/bloodmoon/BloodMoonModel.java
@@ -1,0 +1,4 @@
+package io.novacraft.bloodmoon;
+
+public class BloodMoonModel {
+}
