+++ novacraft-plugin-suite/src/plugin.yml
@@ -13,3 +13,5 @@ commands:
     permission: staffchat.use
     permission-message: "§cYou don't have permission to execute this command!"
     description: What's this?
+  gmp:
+    description: Gives you current moon phase.
